% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drivers.R
\name{computeBasicStats}
\alias{computeBasicStats}
\title{Basic summary statistics of a rooted phylo tree.}
\usage{
computeBasicStats(tree)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted.}
}
\value{
A named vector containing the 16 basic summary statistics.
}
\description{
\code{computeBasicStats} computes basic summary statistics of a rooted binary phylo tree.
}
\seealso{
Other drivers for computing summary statistics: \code{\link{computeNetworkStats}},
  \code{\link{computeSpectralStats}}
}
