% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centralities.R
\name{computeBetweenness}
\alias{computeBetweenness}
\title{Betweenness centrality of a rooted binary phylo tree.}
\usage{
computeBetweenness(tree, weight = FALSE)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted, but the root can be arbitrary.}

\item{weight}{A logical scalar; if TRUE, branch lengths are used; however, they do not change the result!}
}
\description{
\code{computeBetweenness} computes the betweenness centrality of a rooted binary phylo tree.
}
\seealso{
Other network science-based tree shape statistics: \code{\link{computeCloseness}},
  \code{\link{computeDiameter}},
  \code{\link{computeEigenvector}},
  \code{\link{computeFarness}},
  \code{\link{computeWienerIndex}}
}
