% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicStats.R
\name{computeCherries}
\alias{computeCherries}
\title{Compute the number of cherries}
\usage{
computeCherries(tree, DOUBLE = FALSE)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted.}

\item{DOUBLE}{A logical scalar; if TRUE, computes the number of double cherries instead.}
}
\description{
\code{computeCherries} computes the number of cherries of rooted binary phylo tree
}
\seealso{
Other tree shape statistics based on the number of small substructures: \code{\link{computeNum4to8}},
  \code{\link{computePitchforks}}
}
\concept{tree shape statistics based on the number of small substructures}
