% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicStats.R
\name{computeColless}
\alias{computeColless}
\title{Compute the Colless index}
\usage{
computeColless(tree, norm = TRUE)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted.}

\item{norm}{A logical scalar; if TRUE, the value is normalized to lie between 0 and 1.}
}
\description{
\code{computeColless} computes the Colless index of a rooted binary phylo tree
}
\seealso{
Other tree shape statistics based on imbalance: \code{\link{computeSackin}}
}
