% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicStats.R
\name{computeDelW}
\alias{computeDelW}
\title{Compute the maximum delta-width statistic}
\usage{
computeDelW(tree)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted.}
}
\description{
\code{computeMaxDelW} computes the maximum difference of consecutive widths in a rooted binary phylo tree
}
\seealso{
Other tree shape statistics based on height and width: \code{\link{computeMaxHeight}},
  \code{\link{computeMaxWidth}}
}
\concept{tree shape statistics based on height and width}
