% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centralities.R
\name{computeEigenvector}
\alias{computeEigenvector}
\title{Eigenvector centrality of a rooted binary phylo tree.}
\usage{
computeEigenvector(tree, weight = FALSE, scale = FALSE)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted, but the root can be arbitrary.}

\item{weight}{A logical scalar; if TRUE, branch lengths are used; if a root edge is present, it is ignored!}

\item{scale}{A logical scalar; if TRUE, the resulting vector is scaled to have a maximum value of 1.
If no scaling is used (the default), the eigenvector has unit length in the Euclidean norm.}
}
\value{
A list containing the dominant eigenvector and its corresponding eigenvalue.
}
\description{
\code{computeEigenvector} computes the eigenvector centrality of a rooted binary phylo tree.
}
\seealso{
Other network science-based tree shape statistics: \code{\link{computeBetweenness}},
  \code{\link{computeCloseness}},
  \code{\link{computeDiameter}},
  \code{\link{computeFarness}},
  \code{\link{computeWienerIndex}}
}
\concept{network science-based tree shape statistics}
