% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centralities.R
\name{computeFarness}
\alias{computeFarness}
\title{Farness of a rooted binary phylo tree.}
\usage{
computeFarness(tree, weight = FALSE)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted, but the root can be arbitrary.}

\item{weight}{A logical scalar; if TRUE, branch lengths are used; if a root edge is present, it is ignored!}
}
\description{
\code{computeFarness} computes the faress of each node in a rooted binary phylo tree.
}
\seealso{
Other network science-based tree shape statistics: \code{\link{computeBetweenness}},
  \code{\link{computeCloseness}},
  \code{\link{computeDiameter}},
  \code{\link{computeEigenvector}},
  \code{\link{computeWienerIndex}}
}
\concept{network science-based tree shape statistics}
