% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drivers.R
\name{computeLMStats}
\alias{computeLMStats}
\title{Summaries of the distance Lapalcian spectrum (normalized or unnormalized) of a rooted phylo tree.}
\usage{
computeLMStats(tree, norm = FALSE, unitMean = FALSE)
}
\arguments{
\item{tree}{A tree of class \code{phylo}. The tree should be binary and rooted; if it is not, it will be coerced into a rooted tree using \code{root} and/or binarized using \code{multi2di}.}

\item{norm}{A logical vector; a TRUE (FALSE) entry means the Laplacian is (not) normalized.}

\item{unitMean}{A logical scalar; if TRUE, the tree's branch lengths are scaled to unit mean beforehand.}
}
\value{
A named list containing spectral summary statistics: asymmetry, kurtosis, densityMax and lambdaMax.
}
\description{
\code{computeLMStats} computes various summaries of the distance Laplacian specturm of a rooted binary phylo tree.
}
\seealso{
Other drivers for computing summary statistics: \code{\link{computeBasicStats}},
  \code{\link{computeNetworkStats}},
  \code{\link{computeSpectralStats}},
  \code{\link{rankDiscriminatoryStats}}
}
\concept{drivers for computing summary statistics}
