% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra.R
\name{computeMatrix}
\alias{computeMatrix}
\title{Compute the matrix representation of a phylogenetic tree}
\usage{
computeMatrix(tree, weight, dist, full, lap, norm)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted.}

\item{weight}{A logical scalar; if TRUE, branch lengths are used; if a root edge is present, it is ignored!}

\item{dist}{A logical scalar; if TRUE, the distance matrix is computed.}

\item{full}{A logical scalar; if TRUE, the node-node distances are used, otherwise, the leaf-leaf ones; ignored if dist = FALSE.}

\item{lap}{A logical scalar; if TRUE, the Laplacian matrix is computed (defined as diag(rowSums(M))-M for a matrix M).}

\item{norm}{A logical scalar; if TRUE, the Laplacian is a normalized one (i.e. the diagonal is 1); ignored if lap = FALSE.}
}
\description{
\code{computeMatrix} computes a specified matrix representation of a phylogenetic tree
}
\seealso{
Other functions used to compute tree spectra: \code{\link{computeSpectrum}}
}
