% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicStats.R
\name{computeMaxHeight}
\alias{computeMaxHeight}
\title{Compute the maximum height statistic}
\usage{
computeMaxHeight(tree)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted.}
}
\description{
\code{computeMaxHeight} computes the maximum root-to-tip distance in a rooted binary phylo tree
}
\seealso{
Other tree shape statistics based on height and width: \code{\link{computeDelW}},
  \code{\link{computeMaxWidth}}
}
