% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicStats.R
\name{computeMaxWidth}
\alias{computeMaxWidth}
\title{Compute the maximum width statistic}
\usage{
computeMaxWidth(tree)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted.}
}
\description{
\code{computeMaxWidth} computes the maximum number of nodes at a height in a rooted binary phylo tree
}
\seealso{
Other tree shape statistics based on height and width: \code{\link{computeDelW}},
  \code{\link{computeMaxHeight}}
}
\concept{tree shape statistics based on height and width}
