% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicStats.R
\name{computeNum4to8}
\alias{computeNum4to8}
\title{Compute the number of clades of sizes 4 to 8}
\usage{
computeNum4to8(tree)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted.}
}
\value{
A vector of length 5, containing the number of clades of size 4, 5, 6, 7, and 8, in order.
}
\description{
\code{computeNum4to8} computes the number of clades of size 4 to 8 of rooted binary phylo tree
}
\seealso{
Other tree shape statistics based on the number of small substructures: \code{\link{computeCherries}},
  \code{\link{computePitchforks}}
}
\concept{tree shape statistics based on the number of small substructures}
