% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicStats.R
\name{computeSackin}
\alias{computeSackin}
\title{Compute the Sackin index}
\usage{
computeSackin(tree)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted.}
}
\description{
\code{computeSackin} computes the Sackin index of a rooted binary phylo tree
}
\seealso{
Other tree shape statistics based on imbalance: \code{\link{computeColless}}
}
