% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicStats.R
\name{computeStairs1}
\alias{computeStairs1}
\title{Compute the first staircase-ness statistic}
\usage{
computeStairs1(tree)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted.}
}
\description{
\code{computeStairs1} computes the first staircase-ness statistic of a rooted binary phylo tree
}
\seealso{
Other tree shape statistics proposed by Norstrom et al.: \code{\link{computeStairs2}}
}
\concept{tree shape statistics proposed by Norstrom et al.}
