% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicStats.R
\name{computeStairs2}
\alias{computeStairs2}
\title{Compute the second staircase-ness statistic}
\usage{
computeStairs2(tree)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted.}
}
\description{
\code{computeStairs1} computes the second staircase-ness statistic of a rooted binary phylo tree
}
\seealso{
Other tree shape statistics proposed by Norstrom et al.: \code{\link{computeStairs1}}
}
