% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centralities.R
\name{computeWienerIndex}
\alias{computeWienerIndex}
\title{Wiener index of a rooted binary phylo tree.}
\usage{
computeWienerIndex(tree, norm = FALSE, weight = FALSE)
}
\arguments{
\item{tree}{A phylo tree; needs to be binary and rooted, but the root can be arbitrary.}

\item{norm}{A logical scalar; if TRUE, the Wiener index is normalized to give the average path.}

\item{weight}{A logical scalar; if TRUE, branch lengths are used; if a root edge is present, it is ignored!}
}
\description{
\code{computeWienerIndex} computes the Wiener index of a rooted binary phylo tree.
}
\seealso{
Other network science-based tree shape statistics: \code{\link{computeBetweenness}},
  \code{\link{computeCloseness}},
  \code{\link{computeDiameter}},
  \code{\link{computeEigenvector}},
  \code{\link{computeFarness}}
}
