% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drivers.R
\name{rankDiscriminatoryStats}
\alias{rankDiscriminatoryStats}
\title{Ranking of the summary statistics of two lists of trees according to how different the means they produce are.}
\usage{
rankDiscriminatoryStats(tList1, tList2, basic = TRUE, network = TRUE,
  spectral = TRUE, unitMean = TRUE)
}
\arguments{
\item{tList1}{A list of trees of class \code{phylo}. The trees should be binary and rooted; else they are coerced}

\item{tList2}{A list of trees of class \code{phylo}. The trees should be binary and rooted; else they are coerced}

\item{basic}{A logical scalar; if TRUE, computes the basic tree statistics for the lists (lengths must be equal).}

\item{network}{A logical scalar; if TRUE, computes the network statistics for the lists (lengths must be equal).}

\item{spectral}{A logical scalar; if TRUE, computes the spectral statistics for the lists (lengths must be equal).}

\item{unitMean}{A logical scalar; if TRUE, the tree's branch lengths are scaled to unit mean beforehand.}
}
\value{
A named list containing the statistics for the first set of trees, the second set of trees, and p-values.
}
\description{
\code{rankDiscriminatoryStats} ranks different summary stats of rooted binary phylo trees by discriminatory power.
}
\seealso{
Other drivers for computing summary statistics: \code{\link{computeBasicStats}},
  \code{\link{computeLMStats}},
  \code{\link{computeNetworkStats}},
  \code{\link{computeSpectralStats}}
}
\concept{drivers for computing summary statistics}
